Move colorizer.cpp,h to client/utils
